THom from page 28
The endofunctor had to be changed too because it missed
its fmap. And Wikipedia has a nice diagram so
I used it instead. Apparently it's alpha-equivalent
to the article but contains fewer greek letters and
it's not cluttered with projection details and
most important is that there is a diagram.