prove that unitEnc.2 is a prop
propUnitEncCondition is a proof in the style of
coproduct_set that the Path component of unitEnc
is a prop. So it's possible to prove the last
`t` proposition in unitEncWeakEta.

Unfortunately the proof doesn't work with the
type of unitEnc in coproduct_noset, as it's
too polymorphic and carries too little information.

So the next step is to port the recent changes to
a more verbose format of coproduct_noset.